# docs/swagger.yaml
basePath: /
definitions:
  models.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Ben Davis
  description: An example template of a Golang backend API using Fiber and MongoDB
  license:
    name: MIT
  title: The Better Backend Template
  version: "0.1"
paths:
...
  /users:
    get:
      consumes:
      - '*/*'
      description: fetch all users
      parameters:
      - description: Page to seek for in users
        in: query
        name: page
        required: true
        type: integer
      - description: How many items per page
        in: query
        name: take
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Get All Users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: create an user
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Create User
      tags:
      - users
...
swagger: "2.0"